+++ src/ur_rest_node.py
@@ -1,24 +1,20 @@
 """REST-based node for UR robots"""
 
-from enum import Enum
 import traceback
 from typing import Optional, Union
 
-from madsci.common.types.action_types import ActionFailed, ActionSucceeded
+from madsci.client.resource_client import ResourceClient
+from madsci.common.types.action_types import ActionFailed
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.location_types import LocationArgument
 from madsci.common.types.node_types import RestNodeConfig
 from madsci.common.types.resource_types import Pool, Slot
 from madsci.node_module.helpers import action
 from madsci.node_module.rest_node_module import RestNode
+from pydantic import AnyUrl
 from typing_extensions import Annotated
 
-from ur_interface.ur_controller import UR
-from ur_interface.ur_error_types import GripperError, URMovementError
-from ur_interface.ur_kinematics import get_pose_from_joint_angles
-from ur_interface.ur_tools.gripper_controller import FingerGripperController
-from ur_interface.integrated_controller import UREndEffector, IntegratedController
-
+from ur_interface.integrated_controller import IntegratedController, UREndEffector
 
 
 class URNodeConfig(RestNodeConfig):
@@ -28,47 +24,41 @@ class URNodeConfig(RestNodeConfig):
     tcp_pose: list = [0, 0, 0, 0, 0, 0]
     base_reference_frame: Optional[list] = None
     ur_model: str = "UR5e"
-    use_resources: bool = True
     end_effector: Optional[UREndEffector] = None
+    resource_manager_url: Optional[AnyUrl] = None
 
 
 class URNode(RestNode):
     """A Rest Node object to control UR robots"""
 
-    ur_interface: UR = None
+    integrated_controller: IntegratedController = None
     config: URNodeConfig = URNodeConfig()
     config_model = URNodeConfig
 
     def startup_handler(self) -> None:
         """Called to (re)initialize the node. Should be used to open connections to devices or initialize any other resources."""
-        try:
-            # Create templates
-            if self.config.use_resources:
-                self._create_ur_resources()
-            self.create_end_effector_controller()
-
-            self.logger.log("Node initializing...")
-            self.ur_interface = IntegratedController(
-                hostname=self.config.ur_ip,
-                resource_client=self.resource_client if self.config.use_resources else None,
-                tcp_pose=self.config.tcp_pose,
-                base_reference_frame=self.config.base_reference_frame,
-                end_effector=self.config.end_effector,
-                logger=self.logger
-            )
-            self.tool_resource = None
-            self.current_location = None
-        except Exception as err:
-            self.logger.log_error(f"Error starting the UR Node: {err}\n{traceback.format_exc()}")
-            self.startup_has_run = False
-        else:
-            self.startup_has_run = True
-            self.logger.log("UR node initialized!")
+        self.resource_client = None
+        if self.config.resource_manager_url is not None:
+            self.resource_client = ResourceClient(self.config.resource_manager_url)
+            self._create_ur_resources()
+        self.create_end_effector_controller()
+        self.logger.log("Node initializing...")
+        self.integrated_controller = IntegratedController(
+            hostname=self.config.ur_ip,
+            resource_client=self.resource_client if self.config.use_resources else None,
+            tcp_pose=self.config.tcp_pose,
+            base_reference_frame=self.config.base_reference_frame,
+            end_effector=self.config.end_effector,
+            logger=self.logger,
+        )
+        self.tool_resource = None
+        self.current_location = None
+        self.startup_has_run = True
+        self.logger.log("UR node initialized!")
 
     def _create_ur_resources(self) -> None:
         """Create all UR-specific resource templates."""
         if self.config.end_effector == UREndEffector.Robotiq2FingerGripper:
-            
             gripper_slot = Slot(
                 resource_name="robotiq_finger_gripper",
                 resource_class="URGripper",
@@ -127,16 +117,15 @@ class URNode(RestNode):
                 resource_name=f"ur_pipette_{self.node_definition.node_name}",
                 add_to_database=True,
             )
-    
 
     def shutdown_handler(self) -> None:
         """Called to shutdown the node. Should be used to close connections to devices or release any other resources."""
         try:
             self.logger.log("Shutting down")
-            self.ur_interface.disconnect()
+            self.integrated_controller.ur_controller.disconnect()
             self.shutdown_has_run = True
-            del self.ur_interface
-            self.ur_interface = None
+            del self.integrated_controller
+            self.integrated_controller = None
             self.logger.log("Shutdown complete.")
         except Exception as err:
             self.logger.log_error(f"Error shutting down the UR Node: {err}")
@@ -144,21 +133,21 @@ class URNode(RestNode):
     def status_handler(self):
         """Periodically called to update the current status of the node."""
         if not self.node_status.busy:
-            if self.ur_interface:
+            if self.integrated_controller:
                 # Getting robot state
-                self.ur_interface.ur_dashboard.get_overall_robot_status()
-                movement_state, self.current_location = self.ur_interface.get_movement_state()
+                self.integrated_controller.ur_dashboard.get_overall_robot_status()
+                movement_state, self.current_location = self.integrated_controller.ur_controller()
             else:
                 self.logger.log_error("UR interface is not initialized")
                 return
 
-            if "PROTECTIVE_STOP" in self.ur_interface.ur_dashboard.safety_status:
+            if "PROTECTIVE_STOP" in self.integrated_controller.ur_dashboard.safety_status:
                 self.node_status.stopped = True
                 self.logger.log_error("UR is in PROTECTIVE_STOP")
 
-            if "NORMAL" not in self.ur_interface.ur_dashboard.safety_status:
+            if "NORMAL" not in self.integrated_controller.ur_dashboard.safety_status:
                 self.node_status.errored = True
-                self.logger.log_error(f"UR ERROR: {self.ur_interface.ur_dashboard.safety_status}")
+                self.logger.log_error(f"UR ERROR: {self.integrated_controller.ur_dashboard.safety_status}")
             else:
                 self.node_status.errored = False
                 self.node_status.stopped = False
@@ -174,105 +163,28 @@ class URNode(RestNode):
 
     def state_handler(self) -> None:
         """Periodically called to update the current state of the node."""
-        if self.ur_interface:
+        if self.integrated_controller:
             self.node_state = {
                 "current_joint_angles": self.current_location,
             }
 
-    @action(name="getj", description="Get joint angles")
-    def getj(self):
-        """Get joint positions"""
-        joints = self.ur_interface.ur_connection.getj()
-        self.logger.log_info(joints)
-        return ActionSucceeded(data={"joints": joints})
-
-    @action(name="getl", description="Get linear positions")
-    def getl(self):
-        """Get linear position"""
-        lin_pos = self.ur_interface.ur_connection.getl()
-        self.logger.log_info(lin_pos)
-        return ActionSucceeded(data={"lin_pos": lin_pos})
-
-    @action(name="set_freedrive", description="Free robot joints")
-    def set_freedrive(self, timeout: Annotated[int, "how long to do freedrive"] = 60):
-        """set the robot into freedrive"""
-        self.ur_interface.ur_connection.set_freedrive(True, timeout)
-
-    @action(name="set_movement_params", description="Set speed and acceleration parameters")
-    def set_movement_params(
-        self,
-        tcp_pose: Optional[list] = None,
-        velocity: Optional[float] = None,
-        acceleration: Optional[float] = None,
-        gripper_speed: Optional[int] = None,
-        gripper_force: Optional[int] = None,
-    ):
-        """Configure the robot's movement parameters for subsequent transfers"""
-        if tcp_pose is not None:
-            self.ur_interface.ur_connection.set_tcp(tcp_pose)
-        if velocity is not None:
-            self.ur_interface.velocity = velocity
-        if acceleration is not None:
-            self.ur_interface.acceleration = acceleration
-        if gripper_speed is not None:
-            self.ur_interface.gripper_speed = gripper_speed
-        if gripper_force is not None:
-            self.ur_interface.gripper_force = gripper_force
-
-    @action(name="movej", description="Move the robot using joint angles")
-    def movej(
-        self,
-        joints: Annotated[LocationArgument, "Joint angles to move to"],
-        acceleration: Annotated[Optional[float], "Acceleration"] = 0.6,
-        velocity: Annotated[Optional[float], "Velocity"] = 0.6,
-    ):
-        """Move the robot using joint angles"""
-        try:
-            self.logger.log(f"Move joints: {joints.representation}")
-            self.ur_interface.ur_connection.movej(joints=joints.representation, acc=acceleration, vel=velocity)
-
-        except Exception as err:
-            self.logger.log_error(err)
-
-    @action(name="movel", description="Move the robot using linar motion")
-    def movel(
+    @action()
+    def move(
         self,
         target: Annotated[LocationArgument, "Linear location to move to"],
         acceleration: Annotated[Optional[float], "Acceleration"] = 0.6,
         velocity: Annotated[Optional[float], "Velocity"] = 0.6,
-        joint_angle_locations: Annotated[bool, "Use joint angles for the location"] = True,
+        linear_motion: Annotated[bool, "Use linear motion"] = True,
     ):
-        """Move the robot using linear motion"""
-        try:
-            self.logger.log(f"Move location: {target.representation}")
-            if joint_angle_locations:
-                target.representation = get_pose_from_joint_angles(target.representation)
-            self.ur_interface.ur_connection.movel(tpose=target.representation, acc=acceleration, vel=velocity)
+        """Move the robot to target location"""
 
-        except Exception as err:
-            self.logger.log_error(err)
-
-    @action(name="toggle_gripper", description="Move the robot gripper")
-    def toggle_gripper(
-        self,
-        open: Annotated[bool, "Open?"] = False,
-        close: Annotated[bool, "Close?"] = False,
-    ):
-        """Open or close the robot gripper."""
-        gripper = FingerGripperController(hostname=self.config.ur_ip, ur=self.ur_interface, logger=self.logger)
-        self.logger.info("Toggle gripper: Connecting to gripper...")
-        gripper.connect_gripper()
-        self.logger.info("Toggle gripper: Gripper connected")
-        if open:
-            gripper.open_gripper()
-            self.logger.info("Toggle gripper: Gripper opened")
-        elif close:
-            gripper.close_gripper()
-            self.logger.info("Toggle gripper: Gripper closed")
-        else:
-            self.logger.info("Toggle gripper: No action taken")
-        gripper.disconnect_gripper()
-        self.logger.info("Toggle gripper: Gripper disconnected")
+        self.logger.log(f"Move location: {target.representation}")
+        self.integrated_controller.ur_controller.move_to_location(
+            target=target,
+            acceleration=acceleration,
+            velocity=velocity,
+            linear_motion=linear_motion,
+        )
 
     @action(
         name="gripper_transfer",
@@ -289,59 +201,13 @@ class URNode(RestNode):
         target_approach_distance: Annotated[Optional[float], "Approach distance in meters"] = 0.05,
         gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
         gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
     ):
         """Make a transfer using the finger gripper. This function uses linear motions to perform the pick and place movements."""
-        try:
-            if (
-                self.config.use_resources
-                and isinstance(source, LocationArgument)
-                and isinstance(target, LocationArgument)
-            ):
-                self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-                source_resource = self.resource_client.get_resource(source.resource_id)
-                target_resource = self.resource_client.get_resource(target.resource_id)
-                if source_resource.quantity == 0:
-                    raise ValueError("Resource manager: Source location is empty!")
-                if target_resource.quantity != 0:
-                    raise ValueError("Resource manager: Target is occupied!")
 
-            if joint_angle_locations and isinstance(source, LocationArgument) and isinstance(target, LocationArgument):
-                source.representation = get_pose_from_joint_angles(
-                    joints=source.representation, robot_model=self.config.ur_model
-                )
-                target.representation = get_pose_from_joint_angles(
-                    joints=target.representation, robot_model=self.config.ur_model
-                )
-            elif joint_angle_locations and isinstance(source, list) and isinstance(target, list):
-                source = get_pose_from_joint_angles(joints=source, robot_model=self.config.ur_model)
-                target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
+        self.gripper_pick(home, source, source_approach_axis, source_approach_distance, gripper_close)
+        self.gripper_place(home, target, target_approach_axis, target_approach_distance, gripper_open)
 
-            self.ur_interface.gripper_transfer(
-                home=home,
-                source=source,
-                target=target,
-                source_approach_distance=source_approach_distance,
-                target_approach_distance=target_approach_distance,
-                source_approach_axis=source_approach_axis,
-                target_approach_axis=target_approach_axis,
-                gripper_open=gripper_open,
-                gripper_close=gripper_close,
-            )
-
-        except GripperError as err:
-            self.logger.log_error(f"Gripper error during transfer: {err}")
-            return ActionFailed(errors=str(err))
-
-        except URMovementError as err:
-            self.logger.log_error(f"Movement error during transfer: {err}")
-            return ActionFailed(errors=str(err))
-
-        except Exception as err:
-            self.logger.log_error(f"Unexpected error during gripper transfer: {err}")
-            return ActionFailed(errors=str(err))
-
-    @action()
+    @action
     def gripper_pick(
         self,
         home: Annotated[LocationArgument, "Home location"],
@@ -349,44 +215,26 @@ class URNode(RestNode):
         source_approach_axis: Annotated[Optional[str], "Source location approach axis, (X/Y/Z)"] = "z",
         source_approach_distance: Annotated[Optional[float], "Approach distance in meters"] = 0.05,
         gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
     ):
         """Use the gripper to pick a piece of labware from the specified source"""
+        self.logger.log_info(f"Picking from source: {source.representation}")
+
+        self.integrated_controller.gripper_pick(
+            home=home,
+            source=source,
+            source_approach_distance=source_approach_distance,
+            source_approach_axis=source_approach_axis,
+            gripper_close=gripper_close,
+        )
         try:
-            if self.config.use_resources:
-                self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-                source_resource = self.resource_client.get_resource(source.resource_id)
-                if source_resource.quantity == 0:
-                    raise ValueError("Resource manager: Source location is empty!")
+            if self.resource_client is not None:
+                object, _ = self.resource_client.pop(source.resource_id)
+                self.resource_client.push(self.gripper_resource, object)
+        except Exception as e:
+            self.logger.log_error(f"Error during gripper pick: {e}")
+            return ActionFailed(error=traceback.format_exc())
 
-            if joint_angle_locations and isinstance(source, LocationArgument):
-                source.representation = get_pose_from_joint_angles(
-                    joints=source.representation, robot_model=self.config.ur_model
-                )
-            elif joint_angle_locations and isinstance(source, list):
-                source = get_pose_from_joint_angles(joints=source, robot_model=self.config.ur_model)
-            self.logger.log_info(f"Picking from source: {source.representation}")
-            self.ur_interface.gripper_pick(
-                home=home,
-                source=source,
-                source_approach_distance=source_approach_distance,
-                source_approach_axis=source_approach_axis,
-                gripper_close=gripper_close,
-            )
-
-        except GripperError as err:
-            self.logger.log_error(f"Gripper error during pick: {err}")
-            return ActionFailed(errors=str(err))
-
-        except URMovementError as err:
-            self.logger.log_error(f"Movement error during pick: {err}")
-            return ActionFailed(errors=str(err))
-
-        except Exception as err:
-            self.logger.log_error(f"Unexpected error during gripper pick: {err}")
-            return ActionFailed(errors=str(err))
-
-    @action()
+    @action
     def gripper_place(
         self,
         home: Annotated[LocationArgument, "Home location"],
@@ -394,435 +242,44 @@ class URNode(RestNode):
         target_approach_axis: Annotated[Optional[str], "Source location approach axis, (X/Y/Z)"] = "z",
         target_approach_distance: Annotated[Optional[float], "Approach distance in meters"] = 0.05,
         gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
     ):
         """Use the gripper to place a piece of labware at the target."""
-        try:
-            if self.config.use_resources and isinstance(target, LocationArgument):
-                self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-                target_resource = self.resource_client.get_resource(target.resource_id)
-                if target_resource.quantity != 0:
-                    raise ValueError("Resource manager: Target is occupied!")
 
-            if joint_angle_locations and isinstance(target, LocationArgument):
-                target.representation = get_pose_from_joint_angles(
-                    joints=target.representation, robot_model=self.config.ur_model
-                )
-            elif joint_angle_locations and isinstance(target, list):
-                target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-
-            self.ur_interface.gripper_place(
-                home=home,
-                target=target,
-                target_approach_distance=target_approach_distance,
-                target_approach_axis=target_approach_axis,
-                gripper_open=gripper_open,
-            )
-        except GripperError as err:
-            self.logger.log_error(f"Gripper error during place: {err}\n{traceback.format_exc()}")
-            return ActionFailed(errors=str(err))
-
-        except URMovementError as err:
-            self.logger.log_error(f"Movement error during place: {err}\n{traceback.format_exc()}")
-            return ActionFailed(errors=str(err))
-
-        except Exception as err:
-            self.logger.log_error(f"Unexpected error during gripper place: {err}\n{traceback.format_exc()}")
-            return ActionFailed(errors=str(err))
-
-    @action(
-        name="pick_tool",
-        description="Picks up a tool using the provided tool location",
-    )
-    def pick_tool(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        tool_loc: Annotated[Union[LocationArgument, list], "Tool location"],
-        docking_axis: Annotated[Optional[str], "Docking axis, (X/Y/Z)"] = "y",
-        payload: Annotated[Optional[float], "Tool payload"] = None,
-        tool_name: Annotated[Optional[str], "Tool name)"] = None,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Pick a tool with the UR"""
-        if joint_angle_locations and isinstance(tool_loc, LocationArgument):
-            tool_loc.representation = get_pose_from_joint_angles(
-                joints=tool_loc.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(tool_loc, list):
-            tool_loc = get_pose_from_joint_angles(joints=tool_loc, robot_model=self.config.ur_model)
-
-        self.ur_interface.pick_tool(
-            home=home,
-            tool_loc=tool_loc,
-            docking_axis=docking_axis,
-            payload=payload,
-            tool_name=tool_name,
-        )
-
-    @action(name="Place_tool", description="Places the attached tool back to the provided tool docking location")
-    def place_tool(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        tool_docking: Annotated[Union[LocationArgument, list], "Tool docking location"],
-        docking_axis: Annotated[Optional[str], "Docking axis, (X/Y/Z)"] = "y",
-        tool_name: Annotated[Optional[str], "Tool name)"] = None,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Place a tool with the UR"""
-        if joint_angle_locations and isinstance(tool_docking, LocationArgument):
-            tool_docking.representation = get_pose_from_joint_angles(
-                joints=tool_docking.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(tool_docking, list):
-            tool_docking = get_pose_from_joint_angles(joints=tool_docking, robot_model=self.config.ur_model)
-
-        self.ur_interface.place_tool(
-            home=home,
-            tool_loc=tool_docking,
-            docking_axis=docking_axis,
-            tool_name=tool_name,
-        )
-
-    @action(
-        name="gripper_screw_transfer",
-        description="Performs a screw transfer using the Robotiq gripper and custom screwdriving bits",
-    )
-    def gripper_screw_transfer(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        screwdriver_loc: Annotated[Union[LocationArgument, list], "Screwdriver location"],
-        screw_loc: Annotated[Union[LocationArgument, list], "Screw location"],
-        target: Annotated[Union[LocationArgument, list], "Location where the srewdriving will be performed"],
-        screw_time: Annotated[Optional[int], "Srew time in seconds"] = 9,
-        gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
-        gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Make a screwdriving transfer using Robotiq gripper and custom screwdriving bits with UR"""
-
-        if joint_angle_locations and isinstance(screwdriver_loc, LocationArgument):
-            screwdriver_loc.representation = get_pose_from_joint_angles(
-                joints=screwdriver_loc.representation, robot_model=self.config.ur_model
-            )
-            screw_loc.representation = get_pose_from_joint_angles(
-                joints=screw_loc.representation, robot_model=self.config.ur_model
-            )
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(screwdriver_loc, list):
-            screwdriver_loc = get_pose_from_joint_angles(joints=screwdriver_loc, robot_model=self.config.ur_model)
-            screw_loc = get_pose_from_joint_angles(joints=screw_loc, robot_model=self.config.ur_model)
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-
-        self.ur_interface.gripper_screw_transfer(
-            home=home,
-            screwdriver_loc=screwdriver_loc,
-            screw_loc=screw_loc,
-            screw_time=screw_time,
-            target=target,
-            gripper_open=gripper_open,
-            gripper_close=gripper_close,
-        )
-
-    @action(
-        name="pipette_transfer",
-        description="Make a pipette transfer to transfer sample liquids in between two locations",
-    )
-    def pipette_transfer(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        source: Annotated[Union[LocationArgument, list], "Initial location of the sample"],
-        target: Annotated[Union[LocationArgument, list], "Target location of the sample"],
-        tip_loc: Annotated[Union[LocationArgument, list], "New tip location"],
-        tip_trash: Annotated[Union[LocationArgument, list], "Tip trash location"],
-        volume: Annotated[float, "Set a volume in micro liters"],
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Make a pipette transfer for the defined volume with UR"""
-
-        if joint_angle_locations and isinstance(source, LocationArgument):
-            source.representation = get_pose_from_joint_angles(
-                joints=source.representation, robot_model=self.config.ur_model
-            )
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-            tip_loc.representation = get_pose_from_joint_angles(
-                joints=tip_loc.representation, robot_model=self.config.ur_model
-            )
-            tip_trash.representation = get_pose_from_joint_angles(
-                joints=tip_trash.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(source, list):
-            source = get_pose_from_joint_angles(joints=source, robot_model=self.config.ur_model)
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-            tip_loc = get_pose_from_joint_angles(joints=tip_loc, robot_model=self.config.ur_model)
-            tip_trash = get_pose_from_joint_angles(joints=tip_trash, robot_model=self.config.ur_model)
-
-        self.ur_interface.pipette_transfer(
-            home=home,
-            tip_loc=tip_loc,
-            tip_trash=tip_trash,
-            source=source,
-            target=target,
-            volume=volume,
-        )
-
-    @action(
-        name="pipette_pick_and_move_sample",
-        description="Picks and moves a sample using the pipette",
-    )
-    def pipette_pick_and_move_sample(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location in joint angles"],
-        sample_loc: Annotated[Union[LocationArgument, list], "Sample location"],
-        target: Annotated[Union[LocationArgument, list], "Location of the object"],
-        volume: Annotated[int, "Set a volume in micro liters"] = 10,
-        safe_waypoint: Annotated[Union[LocationArgument, list], "Safe waypoint in joint angles"] = None,
-        tip_loc: Annotated[Union[LocationArgument, list], "Tip location"] = None,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-        pipette_speed: Annotated[Optional[int], "Pipette speed in m/s"] = 150,
-    ):
-        """Picks and moves a sample with UR"""
-
-        if joint_angle_locations and isinstance(target, LocationArgument):
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-            sample_loc.representation = get_pose_from_joint_angles(
-                joints=sample_loc.representation, robot_model=self.config.ur_model
-            )
-            tip_loc.representation = (
-                get_pose_from_joint_angles(joints=tip_loc.representation, robot_model=self.config.ur_model)
-                if tip_loc
-                else None
-            )
-        elif joint_angle_locations and isinstance(target, list):
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-            sample_loc = get_pose_from_joint_angles(joints=sample_loc, robot_model=self.config.ur_model)
-            tip_loc = get_pose_from_joint_angles(joints=tip_loc, robot_model=self.config.ur_model) if tip_loc else None
-
-        self.ur_interface.tool_resource_id = self.pipette_resource.resource_id
-
-        self.ur_interface.pipette_pick_and_move_sample(
-            home=home,
-            safe_waypoint=safe_waypoint,
-            sample_loc=sample_loc,
-            target=target,
-            volume=volume,
-            tip_loc=tip_loc,
-            pipette_speed=pipette_speed,
-        )
-
-    @action(
-        name="pipette_dispense_and_retrieve",
-        description="Dispenses a sample and retrieves the pipette tip",
-    )
-    def pipette_dispense_and_retrieve(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location in joint angles"],
-        target: Annotated[Union[LocationArgument, list], "Location of the object"],
-        volume: Annotated[int, "Set a volume in micro liters"] = 10,
-        safe_waypoint: Annotated[Union[LocationArgument, list], "Safe waypoint in joint angles"] = None,
-        tip_trash: Annotated[Union[LocationArgument, list], "Tip trash location"] = None,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-        pipette_speed: Annotated[Optional[int], "Pipette speed in m/s"] = 150,
-    ):
-        """Dispenses a sample and retrieves the pipette with UR"""
-        if joint_angle_locations and isinstance(target, LocationArgument):
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-            tip_trash.representation = (
-                get_pose_from_joint_angles(joints=tip_trash.representation, robot_model=self.config.ur_model)
-                if tip_trash
-                else None
-            )
-        elif joint_angle_locations and isinstance(target, list):
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-            tip_trash = (
-                get_pose_from_joint_angles(joints=tip_trash, robot_model=self.config.ur_model) if tip_trash else None
-            )
-
-        self.ur_interface.tool_resource_id = self.pipette_resource.resource_id
-
-        self.ur_interface.pipette_dispense_and_retrieve(
-            home=home,
-            safe_waypoint=safe_waypoint,
-            target=target,
-            volume=volume,
-            tip_trash=tip_trash,
-            pipette_speed=pipette_speed,
-        )
-
-    @action(
-        name="pick_and_flip_object",
-        description="Picks and flips an object 180 degrees",
-    )
-    def pick_and_flip_object(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        target: Annotated[Union[LocationArgument, list], "Location of the object"],
-        approach_axis: Annotated[Optional[str], "Approach axis, (X/Y/Z)"] = "z",
-        target_approach_distance: Annotated[Optional[float], "Approach distance in meters"] = 0.05,
-        gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
-        gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Picks and flips an object 180 degrees with UR"""
-
-        if joint_angle_locations and isinstance(target, LocationArgument):
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(target, list):
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-
-        self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-
-        self.ur_interface.pick_and_flip_object(
+        self.integrated_controller.gripper_place(
             home=home,
             target=target,
-            approach_axis=approach_axis,
             target_approach_distance=target_approach_distance,
+            target_approach_axis=target_approach_axis,
             gripper_open=gripper_open,
-            gripper_close=gripper_close,
         )
+        try:
+            if self.resource_client is not None:
+                object, _ = self.resource_client.pop(self.gripper_resource)
+                self.resource_client.push(target.resource_id, object)
+        except Exception as e:
+            self.logger.log_error(f"Error during gripper place: {e}")
+            return ActionFailed(error=traceback.format_exc())
 
-    @action(
-        name="remove_cap",
-        description="Removes caps from sample vials",
-    )
-    def remove_cap(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        source: Annotated[Union[LocationArgument, list], "Location of the vial cap"],
-        target: Annotated[
-            Union[LocationArgument, list],
-            "Location of where the cap will be placed after it is removed from the vail",
-        ],
-        gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
-        gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Remove caps from sample vials with UR"""
-        if joint_angle_locations and isinstance(source, LocationArgument):
-            source.representation = get_pose_from_joint_angles(
-                joints=source.representation, robot_model=self.config.ur_model
-            )
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(source, list):
-            source = get_pose_from_joint_angles(joints=source, robot_model=self.config.ur_model)
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-
-        self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-
-        self.ur_interface.remove_cap(
-            home=home,
-            source=source,
-            target=target,
-            gripper_open=gripper_open,
-            gripper_close=gripper_close,
-        )
-
-    @action(
-        name="place_cap",
-        description="Places caps back to sample vials",
-    )
-    def place_cap(
-        self,
-        home: Annotated[Union[LocationArgument, list], "Home location"],
-        source: Annotated[Union[LocationArgument, list], "Vail cap initial location"],
-        target: Annotated[Union[LocationArgument, list], "The vail location where the cap will installed"],
-        gripper_open: Annotated[Optional[int], "Set a max value for the gripper open state"] = 0,
-        gripper_close: Annotated[Optional[int], "Set a min value for the gripper close state"] = 255,
-        joint_angle_locations: Annotated[bool, "Use joint angles for all the locations"] = True,
-    ):
-        """Places caps back to sample vials with UR"""
-        if joint_angle_locations and isinstance(source, LocationArgument):
-            source.representation = get_pose_from_joint_angles(
-                joints=source.representation, robot_model=self.config.ur_model
-            )
-            target.representation = get_pose_from_joint_angles(
-                joints=target.representation, robot_model=self.config.ur_model
-            )
-        elif joint_angle_locations and isinstance(source, list):
-            source = get_pose_from_joint_angles(joints=source, robot_model=self.config.ur_model)
-            target = get_pose_from_joint_angles(joints=target, robot_model=self.config.ur_model)
-
-        self.ur_interface.tool_resource_id = self.gripper_resource.resource_id
-
-        self.ur_interface.place_cap(
-            home=home,
-            source=source,
-            target=target,
-            gripper_open=gripper_open,
-            gripper_close=gripper_close,
-        )
-
-    @action(
-        name="run_urp_program",
-        description="Runs a URP program on the UR",
-    )
-    def run_urp_program(
-        self,
-        transfer_file_path=Annotated[str, "Transfer file path"],
-        program_name=Annotated[str, "Program name"],
-    ):
-        """Run an URP program on the UR"""
-        self.ur_interface.run_urp_program(
-            transfer_file_path=transfer_file_path,
-            program_name=program_name,
-        )
-
-    @action(
-        name="set_digital_io",
-        description="Sets a channel IO output on the UR",
-    )
-    def set_digital_io(
-        self,
-        channel=Annotated[int, "Channel number"],
-        value=Annotated[bool, "True/False"],
-    ):
-        """Sets a channel IO output on the UR"""
-        self.ur_interface.set_digital_io(channel=channel, value=value)
-
-    def get_location(self) -> AdminCommandResponse:
-        """Return the current position of the ur robot"""
-        return AdminCommandResponse(data={"joint_angles": self.ur_interface.ur_connection.getj(), "linear_coordinates": self.ur_interface.getl()})
+    @action(name="e_stop", description="Emergency stop the UR robot")
+    def e_stop(self):
+        """Emergency stop the UR robot"""
+        self.integrated_controller.ur_dashboard.power_off()
+        self.logger.log_info("EMERGENCY STOP EXECUTED")
 
     def reset(self) -> AdminCommandResponse:
         """Reset the ur robot"""
         self.logger.log("Resetting node...")
         # If resetting startup handler does not work, try re-initializing the dashboard
-        self.ur_interface.disconnect()
+        self.integrated_controller.disconnect()
         result = super().reset()
         self.logger.log("Node reset.")
         return result
 
-    @action(name="e_stop", description="Emergency stop the UR robot")
-    def e_stop(self):
-        """Emergency stop the UR robot"""
-        try:
-            self.ur_interface.ur_dashboard.power_off()
-            self.logger.log_info("EMERGENCY STOP EXECUTED")
-
-        except Exception as err:
-            self.logger.log_error(f"FAILED EMERGENCY STOP: {err}")
-            return ActionFailed(errors=str(err))
-
     def safety_stop(self) -> AdminCommandResponse:
         """Safety stop the UR robot"""
-        try:
-            self.ur_interface.ur_dashboard.stop_program()
-            self.logger.log_info("SAFETY STOP EXECUTED")
-            return AdminCommandResponse(success=True)
-        except Exception as err:
-            self.logger.log_error(f"FAILED SAFETY STOP: {err}")
-            return AdminCommandResponse(success=False, errors=str(err))
+        self.integrated_controller.ur_dashboard.stop_program()
+        self.logger.log_info("SAFETY STOP EXECUTED")
+        return AdminCommandResponse(success=True)
 
 
 if __name__ == "__main__":
